--- cli/build_redshift2.go
@@ -1,7 +1,0 @@
-// +build redshift2
-
-package main
-
-import (
-	_ "github.com/golang-migrate/migrate/v4/database/redshift2"
-)
